+++ src/ViewHoldingsRecord.js
@@ -147,12 +147,15 @@ class ViewHoldingsRecord extends React.Component {
     Promise.all([holdingsRecordPromise, instances1Promise])
       .then(([, instances1Response]) => {
         this.setInstance(instances1Response);
-
-        if (this.state.instance?.source) {
-          if (this.isMARCSource() && !this.state.marcRecord) {
-            this.getMARCRecord();
-          }
+        if (!instances1Response.source) {
+          return;
         }
+
+        if (!this.isMARCSource() || this.state.marcRecord) {
+          return;
+        }
+
+        this.getMARCRecord();
       });
   }
 
@@ -1072,15 +1075,22 @@ class ViewHoldingsRecord extends React.Component {
                           }}
                           formatter={{
                             'URL relationship': x => this.refLookup(referenceTables.electronicAccessRelationships, get(x, ['relationshipId'])).name || noValue,
-                            'URI': x => (get(x, ['uri'])
-                              ? (
-                                <a
-                                  href={get(x, ['uri'])}
-                                  style={wrappingCell}
-                                >
-                                  {get(x, ['uri'])}
-                                </a>)
-                              : noValue),
+                            'URI': x => {
+                              const uri = x?.uri;
+
+                              return uri
+                                ? (
+                                  <a
+                                    href={uri}
+                                    rel="noreferrer noopener"
+                                    target="_blank"
+                                    style={wrappingCell}
+                                  >
+                                    {uri}
+                                  </a>
+                                )
+                                : noValue;
+                            },
                             'Link text': x => get(x, ['linkText']) || noValue,
                             'Materials specified': x => get(x, ['materialsSpecification']) || noValue,
                             'URL public note': x => get(x, ['publicNote']) || noValue,