+++ src/providers/DataProviders.test.js
@@ -1,6 +1,9 @@
 import React from 'react';
+
 import { render, screen } from '@folio/jest-config-stripes/testing-library/react';
+
 import { resources, resources2 } from '../../test/fixtures/DataProviders';
+import Harness from '../../test/jest/helpers/Harness';
 import '../../test/jest/__mock__';
 
 import DataProvider from './DataProvider';
@@ -10,13 +13,23 @@ const Children = () => (
 );
 const data = resources;
 const data2 = resources2;
+
+const renderDataProvider = (props) => render(
+  <Harness translations={[]}>
+    <DataProvider {...props}><Children /></DataProvider>
+  </Harness>
+);
+
 describe('DataProvider', () => {
   it('Component should render properly', () => {
-    render(<DataProvider resources={data}><Children /></DataProvider>);
+    renderDataProvider({ resources: data });
+
     expect(screen.getByText('DataProvider')).toBeInTheDocument();
   });
+
   it('Component should be empty', () => {
-    const { container } = render(<DataProvider resources={data2}><Children /></DataProvider>);
+    const { container } = renderDataProvider({ resources: data2 });
+
     expect(container).toBeEmptyDOMElement();
   });
 });