+++ src/settings/NatureOfContentTermsSettings.js
@@ -3,7 +3,10 @@ import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 
 import { ControlledVocab } from '@folio/stripes/smart-components';
-import { IntlConsumer } from '@folio/stripes/core';
+import {
+  IntlConsumer,
+  TitleManager,
+} from '@folio/stripes/core';
 import { getSourceSuppressor } from '@folio/stripes/util';
 
 import { RECORD_SOURCE } from '../constants';
@@ -32,27 +35,32 @@ class NatureOfContentTermsSettings extends React.Component {
     return (
       <IntlConsumer>
         {intl => (
-          <this.connectedControlledVocab
-            {...this.props}
-            baseUrl="nature-of-content-terms"
-            records="natureOfContentTerms"
-            label={<FormattedMessage id="ui-inventory.natureOfContentTerms" />}
-            labelSingular={intl.formatMessage({ id: 'ui-inventory.natureOfContentTerm' })}
-            objectLabel={<FormattedMessage id="ui-inventory.natureOfContentTerms" />}
-            visibleFields={['name', 'source']}
-            columnMapping={{
-              name: intl.formatMessage({ id: 'ui-inventory.name' }),
-              source: intl.formatMessage({ id: 'ui-inventory.source' }),
-            }}
-            readOnlyFields={['source']}
-            itemTemplate={{ source: 'local' }}
-            hiddenFields={['description', 'numberOfObjects']}
-            actionSuppressor={{ edit: suppress, delete: suppress }}
-            nameKey="name"
-            id="natureOfContentTerms"
-            sortby="name"
-            editable={hasPerm}
-          />
+          <TitleManager
+            page={intl.formatMessage({ id: 'ui-inventory.settings.inventory.title' })}
+            record={intl.formatMessage({ id: 'ui-inventory.natureOfContentTerms' })}
+          >
+            <this.connectedControlledVocab
+              {...this.props}
+              baseUrl="nature-of-content-terms"
+              records="natureOfContentTerms"
+              label={<FormattedMessage id="ui-inventory.natureOfContentTerms" />}
+              labelSingular={intl.formatMessage({ id: 'ui-inventory.natureOfContentTerm' })}
+              objectLabel={<FormattedMessage id="ui-inventory.natureOfContentTerms" />}
+              visibleFields={['name', 'source']}
+              columnMapping={{
+                name: intl.formatMessage({ id: 'ui-inventory.name' }),
+                source: intl.formatMessage({ id: 'ui-inventory.source' }),
+              }}
+              readOnlyFields={['source']}
+              itemTemplate={{ source: 'local' }}
+              hiddenFields={['description', 'numberOfObjects']}
+              actionSuppressor={{ edit: suppress, delete: suppress }}
+              nameKey="name"
+              id="natureOfContentTerms"
+              sortby="name"
+              editable={hasPerm}
+            />
+          </TitleManager>
         )}
       </IntlConsumer>
     );