+++ src/views/ItemView.js
@@ -805,6 +805,7 @@ class ItemView extends React.Component {
     };
 
     const enumerationData = {
+      displaySummary: get(item, 'displaySummary', '-'),
       enumeration: get(item, 'enumeration', '-'),
       chronology: get(item, 'chronology', '-'),
       volume: get(item, 'volume', '-'),
@@ -905,15 +906,22 @@ class ItemView extends React.Component {
     const electronicAccessFormatter = {
       'URL relationship': x => refLookup(referenceTables.electronicAccessRelationships,
         get(x, ['relationshipId'])).name || noValue,
-      'URI': x => (get(x, ['uri'])
-        ? (
-          <a
-            href={get(x, ['uri'])}
-            style={wrappingCell}
-          >
-            {get(x, ['uri'])}
-          </a>)
-        : noValue),
+      'URI': x => {
+        const uri = x?.uri;
+
+        return uri
+          ? (
+            <a
+              href={uri}
+              rel="noreferrer noopener"
+              target="_blank"
+              style={wrappingCell}
+            >
+              {uri}
+            </a>
+          )
+          : noValue;
+      },
       'Link text': x => get(x, ['linkText']) || noValue,
       'Materials specified': x => get(x, ['materialsSpecification']) || noValue,
       'URL public note': x => get(x, ['publicNote']) || noValue,
@@ -1320,6 +1328,17 @@ class ItemView extends React.Component {
                       id="acc03"
                       label={<FormattedMessage id="ui-inventory.enumerationData" />}
                     >
+                      <Row>
+                        <Col
+                          smOffset={0}
+                          sm={4}
+                        >
+                          <KeyValue
+                            label={<FormattedMessage id="ui-inventory.displaySummary" />}
+                            value={checkIfElementIsEmpty(enumerationData.displaySummary)}
+                          />
+                        </Col>
+                      </Row>
                       <Row>
                         <Col
                           smOffset={0}